+++ rollup.config.js
@@ -1,3 +1,4 @@
+import globals from 'rollup-plugin-node-globals';
 import commonjs from 'rollup-plugin-commonjs';
 import resolve from 'rollup-plugin-node-resolve';
 import babel from 'rollup-plugin-babel';
@@ -9,26 +10,15 @@ const plugins = [
 		exclude:      ['**/*.json', 'node_modules/**'],
 		throwOnError: process.env.ROLLUP_WATCH != 'true'
 	}),
-	babel(Object.assign({
-		runtimeHelpers: true,
-		babelrc:        false,
-		exclude:        'node_modules/**'
-	}, pkg.babel, {
-		presets: pkg.babel.presets.map((preset) => {
-
-			if (Array.isArray(preset) && preset[0] == 'env') {
-				preset[1].modules = false;
-			}
-
-			return preset;
-		})
-	})),
+	babel({
+		runtimeHelpers: true
+	}),
 	resolve({
 		preferBuiltins: true
 	}),
-	commonjs()
+	commonjs(),
+	globals()
 ];
-
 const dependencies = [].concat(
 	['os'],
 	Object.keys(pkg.dependencies)
@@ -42,14 +32,11 @@ function external(id) {
 
 export default {
 	input:  'src/index.js',
-	watch:  {
-		include: 'src/**/*.js'
-	},
 	plugins,
 	external,
 	output: {
 		file:      pkg.main,
 		format:    'cjs',
-		sourcemap: true
+		sourcemap: 'inline'
 	}
 };