+++ src/index.js
@@ -1,6 +1,6 @@
 import { cpus } from 'os';
 import through from 'through2-concurrent';
-import SrcsetGenerator from './generator';
+import SrcsetGenerator, { matchImage } from '@flexis/srcset';
 
 const throughOptions = {
 	maxConcurrency: cpus().length
@@ -15,7 +15,6 @@ export default function plugin(rules = [], inputOptions = {}) {
 		skipOptimization: false,
 		scalingUp:        true
 	}, inputOptions);
-
 	const srcset = new SrcsetGenerator(options);
 
 	async function each(file, enc, next) {
@@ -25,17 +24,21 @@ export default function plugin(rules = [], inputOptions = {}) {
 			return;
 		}
 
-		const push = this.push.bind(this);
-
 		try {
 
 			const results = await Promise.all(
 				rules.map(async (rule) => {
 
-					const matches = await srcset.matchImage(file, rule.match);
+					const matches = await matchImage(file, rule.match);
 
 					if (matches) {
-						await srcset.generate(file, rule, push);
+
+						const images = srcset.generate(file, rule);
+
+						for await (const image of images) {
+							this.push(image);
+						}
+
 						return true;
 					}