+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2018-07-07
+### Added
+- pre-push hook with husky.
+- `npm-check`
+
+### Changed
+- Dependencies are updated.
+- Code-style fixes.
+- `package.babel` -> `.babelrc`.
+- `.eslintrc.js` -> `.eslintrc`.
+- 'src/generator.js' -> `@flexis/srcset`.
+
 ## [1.2.1] - 2018-03-31
 ### Changed
 - Some `devDependencies` are updated.